Fix default vector storage
